+++ CHANGELOG.md
@@ -36,6 +36,8 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Changed
 
+- The mobile and tablet header now stays sticky while scrolling ([#7840]).
+
 - The `edge` channel has been switched to the new UI and versioning scheme.
 
 ### Deprecated
@@ -50,6 +52,7 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 [go-1.26.6]: https://groups.google.com/g/golang-announce/c/94pEornpRlI
 [#7514]: https://github.com/AdguardTeam/AdGuardHome/issues/7514
+[#7840]: https://github.com/AdguardTeam/AdGuardHome/pull/7840
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
 
 <!--